inet: fix flags used when deleting routes
When using PreferredTechnologies, default routes may be added and
deleted multiple times as interfaces of different priorities appear.
Unfortunately on kernel 6.11 connman is successfully adding routes, but
it is receiving "Operation not supported (95)" errors when it tries to
delete those same routes.

strace shows that connman is passing unexpected flags for RTM_DELROUTE:
  nlmsg_flags=NLM_F_REQUEST|NLM_F_ACK|NLM_F_BULK|0x400

connman is actually passing RTM_NEWROUTE's NLM_F_EXCL (0x200) and
NLM_F_CREATE (0x400) which are interpreted as RTM_DELROUTE's NLM_F_BULK
(0x200) and "nothing", respectively.

The NLM_F_BULK flag was added to the kernel in commit
545528d788556c724eeb5400757f828ef27782a8 ("net: netlink: add NLM_F_BULK
delete request modifier") which is included in the 5.19 and 6.0
kernels. This flag was likely ignored in prior kernel versions which
was why route deletions succeeded in the past.

The fix is to ensure these flags are not set for RTM_DELROUTE.